requirements.txt
pandas == 1.4.3
streamlit==1.11.1
Pillow == 9.2.0
numpy ==1.23.1
pivottablejs == 0.9.0
matplotlib == 3.5.2
pydeck ==0.7.1

